data_lake/populate-worker/requirements.txt
python-dotenv
Flask-PyMongo
pymongo[srv]
requests
